+++ logindirector/logindirector.csproj
@@ -30,9 +30,9 @@
     <PackageReference Include="Rollbar.NetCore.AspNet" Version="5.2.0" />
     <PackageReference Include="Amazon.AspNetCore.DataProtection.SSM" Version="3.1.1" />
     <PackageReference Include="Amazon.Extensions.Configuration.SystemsManager" Version="5.1.0" />
-    <PackageReference Include="AWSSDK.Core" Version="3.7.200.6" />
-    <PackageReference Include="AWSSDK.SecurityToken" Version="3.7.201.2" />
-    <PackageReference Include="AWSSDK.SimpleSystemsManagement" Version="3.7.200.6" />
+    <PackageReference Include="AWSSDK.Core" Version="3.7.200.9" />
+    <PackageReference Include="AWSSDK.SecurityToken" Version="3.7.201.5" />
+    <PackageReference Include="AWSSDK.SimpleSystemsManagement" Version="3.7.200.9" />
     <PackageReference Include="Steeltoe.Extensions.Configuration.CloudFoundryCore" Version="3.2.3" />
     <PackageReference Include="MSTest.TestFramework" Version="3.1.1" />
     <PackageReference Include="System.Net.Http" Version="4.3.4" />
